pc/at - we have stats at repo level for project repos for user edits
Next step will be to combine those.
